+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - All parameters on `NewType.__call__` are now positional-only. This means that
   the signature of `typing_extensions.NewType.__call__` now exactly matches the
   signature of `typing.NewType.__call__`. Patch by Alex Waygood.
+- Fix bug with using `@deprecated` on a mixin class. Patch by Jelle Zijlstra.
 
 # Release 4.8.0 (September 17, 2023)